requirements.txt
Flask
scikit-learn
pandas
joblib
google-generativeai
nltk
python-dotenv
gunicorn
numpy
